+++ packages/nimble-components/src/table/components/header/template.ts
@@ -3,7 +3,7 @@ import type { TableHeader } from '.';
 import { iconArrowDownTag } from '../../../icons/arrow-down';
 import { iconArrowUpTag } from '../../../icons/arrow-up';
 import { iconTwoSquaresInBracketsTag } from '../../../icons/two-squares-in-brackets';
-import { TableColumnSortDirection } from '../../types';
+import { TableColumnAlignment, TableColumnSortDirection } from '../../types';
 import {
     tableColumnHeaderGroupedLabel,
     tableColumnHeaderSortedAscendingLabel,
@@ -13,6 +13,7 @@ import {
 // prettier-ignore
 export const template = html<TableHeader>`
     <template role="columnheader"
+        class="${x => (x.alignment === TableColumnAlignment.right ? 'right-align' : '')}"
         aria-sort="${x => x.ariaSort}"
         ${'' /* Prevent header double clicks from selecting text */}
         @mousedown="${(_x, c) => !((c.event as MouseEvent).detail > 1)}"