+++ packages/nimble-components/src/table/components/header/tests/table-header.spec.ts
@@ -2,7 +2,7 @@ import { html } from '@microsoft/fast-element';
 import { TableHeader } from '..';
 import { waitForUpdatesAsync } from '../../../../testing/async-helpers';
 import { type Fixture, fixture } from '../../../../utilities/tests/fixture';
-import { TableColumnSortDirection } from '../../../types';
+import { TableColumnAlignment, TableColumnSortDirection } from '../../../types';
 import { TableHeaderPageObject } from './table-header-pageobject';
 
 async function setup(): Promise<Fixture<TableHeader>> {
@@ -146,5 +146,9 @@ describe('TableHeader', () => {
         expect(element.getAttribute('aria-sort')).toEqual('descending');
         expect(pageObject.isSortAscendingIconVisible()).toBeFalse();
         expect(pageObject.isSortDescendingIconVisible()).toBeFalse();
+    });
+
+    it('defaults to left-aligned', () => {
+        expect(element.alignment).toBe(TableColumnAlignment.left);
     });
 });